+++ packages/tswebextension/src/lib/mv2/background/cosmetic-api.ts
@@ -188,7 +188,15 @@ export class CosmeticApi {
             return undefined;
         }
 
-        return scriptText;
+        return `
+        (function () {
+            try {
+                ${scriptText}
+            } catch (ex) {
+                console.error('Error executing AG js: ' + ex);
+            }
+        })();
+        `;
     }
 
     /**